requirements.txt
# === Backend & Server ===
fastapi>=0.111
uvicorn>=0.30

# === ML stack ===
torch    
transformers>=4.42
numpy>=1.26
datasets>=2.20
pandas>=2.2
python-dotenv>=1.0

# === DB ===
pymysql>=1.1
sqlmodel>=0.0.22

# === LangChain core + providers ===
langchain>=0.2.12
langchain-openai>=0.1.13
openai>=1.40

# Google GenAI
langchain-google-genai>=2.0.0

# === Pinecone v5 ===
pinecone-client>=5.0.1
langchain-pinecone>=0.1.3


aiohttp>=3.9
